updated folder names
