--- components/User/User.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const User = require('./UserModel');
-
-const router = express.Router();
-
-/* GET home page. */
-
-module.exports = router;
